requirements.txt
streamlit
langchain-groq
python-docx
python-dotenv
